(chore): changed endpoint for wadhwani college and zonal leaderboard
