+++ api/leaderboard/serializers.py
@@ -87,3 +87,14 @@ class StudentLeaderboardSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ["full_name", "total_karma", "institution"]
+
+class WadhwaniCollegeLeaderboardSerializer(serializers.Serializer):
+    code = serializers.CharField()
+    title = serializers.CharField()
+    total_karma = serializers.IntegerField()
+    students = serializers.IntegerField()
+
+class WadhwaniZoneLeaderboardSerializer(serializers.Serializer):
+    zone_name = serializers.CharField()
+    total_karma = serializers.IntegerField()
+    students = serializers.IntegerField()
