+++ api/leaderboard/urls.py
@@ -6,5 +6,7 @@ urlpatterns = [
     path('students/', leaderboard_view.StudentsLeaderboard.as_view()),
     path('students-monthly/', leaderboard_view.StudentsMonthlyLeaderboard.as_view()),
     path('college/', leaderboard_view.CollegeLeaderboard.as_view()),
-    path('college-monthly/', leaderboard_view.CollegeMonthlyLeaderboard.as_view())
+    path('college-monthly/', leaderboard_view.CollegeMonthlyLeaderboard.as_view()),
+    path('wadhwani-college/', leaderboard_view.WadhwaniCollegeLeaderboard.as_view()),
+    path('wadhwani-zonal/', leaderboard_view.WadhwaniZonalLeaderboard.as_view()),
 ]
