+++ api/urls.py
@@ -15,6 +15,5 @@ urlpatterns = [
     path('top100/', include('api.top100_coders.urls')),
     path('launchpad/', include('api.launchpad.urls')),
     path('donate/', include('api.donate.urls')),
-    path('wadhwani/', include('api.wadhwani.urls')),
     path("__debug__/", include(debug_toolbar.urls)),
 ]
